fix trim

+++ gdsfactory/component.py
@@ -431,16 +431,18 @@ class ComponentBase:
             right: right coordinate of the bounding box.
             top: top coordinate of the bounding box.
             flatten: if True, flattens the Component.
-
         """
         c = self
-        _kdb_cell = c.kcl.clip(c._kdb_cell, kdb.DBox(left, bottom, right, top))
-        c._kdb_cell.clear()
-        c.copy_tree(_kdb_cell)
-        c.rebuild()
-        _kdb_cell.delete()
-        if flatten:
-            c.flatten()
+
+        domain_box = kdb.DBox(left, bottom, right, top)
+        if not c.dbbox().inside(domain_box):
+            _kdb_cell = c.kcl.clip(c._kdb_cell, kdb.DBox(left, bottom, right, top))
+            c._kdb_cell.clear()
+            c.copy_tree(_kdb_cell)
+            c.rebuild()
+            _kdb_cell.delete()
+            if flatten:
+                c.flatten()
 
     def add_polygon(
         self,
@@ -982,14 +984,21 @@ class ComponentBase:
 
         return get_netlist(self, **kwargs)
 
-    def write_netlist(self, filepath: str, **kwargs) -> pathlib.Path:
-        """Write netlist in YAML."""
-        netlist = self.get_netlist(**kwargs)
+    def write_netlist(
+        self, netlist: dict[str, Any], filepath: str | pathlib.Path | None = None
+    ) -> str:
+        """Returns netlist as YAML string.
+
+        Args:
+            netlist: netlist to write.
+            filepath: Optional file path to write to.
+        """
         netlist = convert_tuples_to_lists(netlist)
-        yaml_component = yaml.dump(netlist)
-        filepath = pathlib.Path(filepath)
-        filepath.write_text(yaml_component)
-        return filepath
+        yaml_string = yaml.dump(netlist)
+        if filepath:
+            filepath = pathlib.Path(filepath)
+            filepath.write_text(yaml_string)
+        return yaml_string
 
     def plot_netlist(
         self,
